Try to fix loading comments with and without SSL
Disqus would treat pages loaded over http and https as separate,
which resulted in split comment threads depending on how people got to
the website. It was also blocking a complete migration to https.

Now we're forcing the thread's url to start with http://blog.ragnarson,
which will force disqus to load the same content for all sources,
including local development machines. I chose http so that we don't
lose previous data.