Add blockchain as a posts tag
The separated category for blockchain related content is needed to
show outside our expertise in this field. With one link we will
be able to show that we are in the field for over one year.

It will be used in our profile on moneo.io - blockchain experts
and clients matching platform. We can also use it on our
blockchain related landing pages